# ground-control/api/v1/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Ground Control management API. Bearer authentication is the recommended workflow for interactive clients. Basic authentication remains supported on authenticated human endpoints for automation compatibility.",
    "title": "Harbor Satellite Ground Control API",
    "version": "1.0"
  },
  "basePath": "/",
  "paths": {
    "/api/groups/sync": {
      "post": {
        "description": "Creates the group described in the state artifact, updates the projects of all robot accounts attached to its satellites, ensures the \"satellite\" project exists in Harbor, and writes the state artifact. Requires an authenticated user. Bearer auth is recommended; Basic auth remains supported for automation.",
        "tags": [
          "groups"
        ],
        "summary": "Create or update a group from a state artifact",
        "operationId": "syncGroup",
        "parameters": [
          {
            "name": "state",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StateArtifact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group created and state synced",
            "schema": {
              "$ref": "#/definitions/Group"
            }
          },
          "400": {
            "description": "Invalid request body",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "502": {
            "description": "Upstream Harbor error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "Artifact": {
      "type": "object",
      "properties": {
        "deleted": {
          "type": "boolean"
        },
        "digest": {
          "type": "string"
        },
        "labels": {
          "description": "Free-form label metadata; passed through unchanged.",
          "type": "object"
        },
        "repository": {
          "type": "string"
        },
        "tag": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "type": {
          "type": "string"
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "required": [
        "error"
      ],
      "properties": {
        "error": {
          "type": "string"
        }
      }
    },
    "Group": {
      "type": "object",
      "required": [
        "id",
        "group_name",
        "registry_url",
        "projects",
        "created_at",
        "updated_at"
      ],
      "properties": {
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "group_name": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "projects": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "registry_url": {
          "type": "string"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "StateArtifact": {
      "type": "object",
      "properties": {
        "artifacts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Artifact"
          }
        },
        "group": {
          "type": "string"
        },
        "registry": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BasicAuth": {
      "type": "basic"
    },
    "BearerAuth": {
      "description": "Use `Bearer \u003ctoken\u003e`.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}